+++ awc/src/middleware/redirect.rs
@@ -190,10 +190,7 @@ where
                         let body_new = if is_redirect {
                             // try to reuse body
                             match body {
-                                Some(ref bytes) => AnyBody::Bytes {
-                                    body: bytes.clone(),
-                                },
-                                // TODO: should this be AnyBody::Empty or AnyBody::None.
+                                Some(ref bytes) => AnyBody::from(bytes.clone()),
                                 _ => AnyBody::empty(),
                             }
                         } else {